Struct aws_sdk_acm::client::fluent_builders::RequestCertificate

pub struct RequestCertificate<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> RequestCertificate<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<RequestCertificateOutput, SdkError<RequestCertificateError>> where
R::Policy: SmithyRetryPolicy<RequestCertificateInputOperationOutputAlias, RequestCertificateOutput, RequestCertificateError, RequestCertificateInputOperationRetryAlias>,

pub fn domain_name(self, inp: impl Into<String>) -> Self

Fully qualified domain name (FQDN), such as www.example.com, that you want to secure with an ACM certificate. Use an asterisk (*) to create a wildcard certificate that protects several sites in the same domain. For example, *.example.com protects www.example.com, site.example.com, and images.example.com.

The first domain name you enter cannot exceed 64 octets, including periods. Each subsequent Subject Alternative Name (SAN), however, can be up to 253 octets in length.

pub fn set_domain_name(self, input: Option<String>) -> Self

pub fn validation_method(self, inp: ValidationMethod) -> Self

The method you want to use if you are requesting a public certificate to validate that you own or control domain. You can validate with DNS or validate with email. We recommend that you use DNS validation.

pub fn set_validation_method(self, input: Option<ValidationMethod>) -> Self

pub fn subject_alternative_names(self, inp: impl Into<String>) -> Self

Appends an item to SubjectAlternativeNames.

To override the contents of this collection use set_subject_alternative_names.

Additional FQDNs to be included in the Subject Alternative Name extension of the ACM certificate. For example, add the name www.example.net to a certificate for which the DomainName field is www.example.com if users can reach your site by using either name. The maximum number of domain names that you can add to an ACM certificate is 100. However, the initial quota is 10 domain names. If you need more than 10 names, you must request a quota increase. For more information, see Quotas.

The maximum length of a SAN DNS name is 253 octets. The name is made up of multiple labels separated by periods. No label can be longer than 63 octets. Consider the following examples:

(63 octets).(63 octets).(63 octets).(61 octets) is legal because the total length is 253 octets (63+1+63+1+63+1+61) and no label exceeds 63 octets.

(64 octets).(63 octets).(63 octets).(61 octets) is not legal because the total length exceeds 253 octets (64+1+63+1+63+1+61) and the first label exceeds 63 octets.

(63 octets).(63 octets).(63 octets).(62 octets) is not legal because the total length of the DNS name (63+1+63+1+63+1+62) exceeds 253 octets.

pub fn set_subject_alternative_names(self, input: Option<Vec<String>>) -> Self

pub fn idempotency_token(self, inp: impl Into<String>) -> Self

Customer chosen string that can be used to distinguish between calls to RequestCertificate. Idempotency tokens time out after one hour. Therefore, if you call RequestCertificate multiple times with the same idempotency token within one hour, ACM recognizes that you are requesting only one certificate and will issue only one. If you change the idempotency token for each call, ACM recognizes that you are requesting multiple certificates.

pub fn set_idempotency_token(self, input: Option<String>) -> Self

pub fn domain_validation_options(
self,
inp: impl Into<DomainValidationOption>
) -> Self

Appends an item to DomainValidationOptions.

To override the contents of this collection use set_domain_validation_options.

The domain name that you want ACM to use to send you emails so that you can validate domain ownership.

pub fn set_domain_validation_options(
self,
input: Option<Vec<DomainValidationOption>>
) -> Self

pub fn options(self, inp: CertificateOptions) -> Self

Currently, you can use this parameter to specify whether to add the certificate to a certificate transparency log. Certificate transparency makes it possible to detect SSL/TLS certificates that have been mistakenly or maliciously issued. Certificates that have not been logged typically produce an error message in a browser. For more information, see Opting Out of Certificate Transparency Logging.

pub fn set_options(self, input: Option<CertificateOptions>) -> Self

pub fn certificate_authority_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the private certificate authority (CA) that will be used to issue the certificate. If you do not provide an ARN and you are trying to request a private certificate, ACM will attempt to issue a public certificate. For more information about private CAs, see the Amazon Web Services Certificate Manager Private Certificate Authority (PCA) user guide. The ARN must have the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

pub fn set_certificate_authority_arn(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

One or more resource tags to associate with the certificate.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self