Module aws_sdk_acm::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

invalid_args_exception

See InvalidArgsException

invalid_arn_exception

See InvalidArnException

invalid_domain_validation_options_exception

See InvalidDomainValidationOptionsException

invalid_parameter_exception

See InvalidParameterException

invalid_state_exception

See InvalidStateException

invalid_tag_exception

See InvalidTagException

limit_exceeded_exception

See LimitExceededException

request_in_progress_exception

See RequestInProgressException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

tag_policy_exception

See TagPolicyException

throttling_exception

See ThrottlingException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have access

AddTagsToCertificateError

ConflictException

You are trying to

DeleteCertificateError

DescribeCertificateError

ExportCertificateError

GetAccountConfigurationError

GetCertificateError

ImportCertificateError

InvalidArgsException

One or more of of request parameters specified is not valid.

InvalidArnException

The requested Amazon Resource Name (ARN) does not refer to an existing resource.

InvalidDomainValidationOptionsException

One or more values in the DomainValidationOption structure is

InvalidParameterException

An input parameter was invalid.

InvalidStateException

Processing has reached an invalid state.

InvalidTagException

One or both of the values that make up the key-value pair is not valid. For example, you

LimitExceededException

An ACM quota has been exceeded.

ListCertificatesError

ListTagsForCertificateError

PutAccountConfigurationError

RemoveTagsFromCertificateError

RenewCertificateError

RequestCertificateError

RequestInProgressException

The certificate request is in process and the certificate in your account has not yet been

ResendValidationEmailError

ResourceInUseException

The certificate is in use by another Amazon Web Services service in the caller's account. Remove the

ResourceNotFoundException

The specified certificate cannot be found in the caller's account or the caller's account

TagPolicyException

A specified tag did not comply with an existing tag policy and was rejected.

ThrottlingException

The request was denied

TooManyTagsException

The request contains too many tags. Try the request again with fewer tags.

UpdateCertificateOptionsError

ValidationException

The supplied input

Enums

AddTagsToCertificateErrorKind

DeleteCertificateErrorKind

DescribeCertificateErrorKind

ExportCertificateErrorKind

GetAccountConfigurationErrorKind

GetCertificateErrorKind

ImportCertificateErrorKind

ListCertificatesErrorKind

ListTagsForCertificateErrorKind

PutAccountConfigurationErrorKind

RemoveTagsFromCertificateErrorKind

RenewCertificateErrorKind

RequestCertificateErrorKind

ResendValidationEmailErrorKind

UpdateCertificateOptionsErrorKind