Module aws_sdk_acm::model

Modules

certificate_detail

See CertificateDetail

certificate_options

See CertificateOptions

certificate_summary

See CertificateSummary

domain_validation

See DomainValidation

domain_validation_option

See DomainValidationOption

expiry_events_configuration

See ExpiryEventsConfiguration

extended_key_usage

See ExtendedKeyUsage

filters

See Filters

key_usage

See KeyUsage

renewal_summary

See RenewalSummary

resource_record

See ResourceRecord

tag

See Tag

Structs

CertificateDetail

Contains metadata about an ACM certificate. This structure is returned in the response

CertificateOptions

Structure that contains options for your certificate. Currently, you can use this only to

CertificateSummary

This structure is returned in the response object of ListCertificates

DomainValidation

Contains information about the validation of each domain name in the certificate.

DomainValidationOption

Contains information about the domain names that you want ACM to use to send you emails

ExpiryEventsConfiguration

Object containing

ExtendedKeyUsage

The Extended Key Usage X.509 v3 extension defines one or more purposes for which the

Filters

This structure can be used in the ListCertificates action to filter the

KeyUsage

The Key Usage X.509 v3 extension defines the purpose of the public key contained in the

RenewalSummary

Contains information about the status of ACM's managed renewal for the certificate. This

ResourceRecord

Contains a DNS record value that you can use to validate ownership or control

Tag

A key-value pair that identifies or specifies metadata about an ACM resource.

Enums

CertificateStatus

CertificateTransparencyLoggingPreference

CertificateType

DomainStatus

ExtendedKeyUsageName

FailureReason

KeyAlgorithm

KeyUsageName

RecordType

RenewalEligibility

RenewalStatus

RevocationReason

ValidationMethod