Struct aws_sdk_acm::model::certificate_detail::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CertificateDetail

Implementations

impl Builder

pub fn certificate_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

pub fn set_certificate_arn(self, input: Option<String>) -> Self

pub fn domain_name(self, input: impl Into<String>) -> Self

The fully qualified domain name for the certificate, such as www.example.com or example.com.

pub fn set_domain_name(self, input: Option<String>) -> Self

pub fn subject_alternative_names(self, input: impl Into<String>) -> Self

pub fn set_subject_alternative_names(self, input: Option<Vec<String>>) -> Self

pub fn domain_validation_options(
self,
input: impl Into<DomainValidation>
) -> Self

pub fn set_domain_validation_options(
self,
input: Option<Vec<DomainValidation>>
) -> Self

pub fn serial(self, input: impl Into<String>) -> Self

The serial number of the certificate.

pub fn set_serial(self, input: Option<String>) -> Self

pub fn subject(self, input: impl Into<String>) -> Self

The name of the entity that is associated with the public key contained in the certificate.

pub fn set_subject(self, input: Option<String>) -> Self

pub fn issuer(self, input: impl Into<String>) -> Self

The name of the certificate authority that issued and signed the certificate.

pub fn set_issuer(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The time at which the certificate was requested.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn issued_at(self, input: Instant) -> Self

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

pub fn set_issued_at(self, input: Option<Instant>) -> Self

pub fn imported_at(self, input: Instant) -> Self

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

pub fn set_imported_at(self, input: Option<Instant>) -> Self

pub fn status(self, input: CertificateStatus) -> Self

The status of the certificate.

pub fn set_status(self, input: Option<CertificateStatus>) -> Self

pub fn revoked_at(self, input: Instant) -> Self

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

pub fn set_revoked_at(self, input: Option<Instant>) -> Self

pub fn revocation_reason(self, input: RevocationReason) -> Self

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

pub fn set_revocation_reason(self, input: Option<RevocationReason>) -> Self

pub fn not_before(self, input: Instant) -> Self

The time before which the certificate is not valid.

pub fn set_not_before(self, input: Option<Instant>) -> Self

pub fn not_after(self, input: Instant) -> Self

The time after which the certificate is not valid.

pub fn set_not_after(self, input: Option<Instant>) -> Self

pub fn key_algorithm(self, input: KeyAlgorithm) -> Self

The algorithm that was used to generate the public-private key pair.

pub fn set_key_algorithm(self, input: Option<KeyAlgorithm>) -> Self

pub fn signature_algorithm(self, input: impl Into<String>) -> Self

The algorithm that was used to sign the certificate.

pub fn set_signature_algorithm(self, input: Option<String>) -> Self

pub fn in_use_by(self, input: impl Into<String>) -> Self

pub fn set_in_use_by(self, input: Option<Vec<String>>) -> Self

pub fn failure_reason(self, input: FailureReason) -> Self

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

pub fn set_failure_reason(self, input: Option<FailureReason>) -> Self

pub fn type(self, input: CertificateType) -> Self

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

pub fn set_type(self, input: Option<CertificateType>) -> Self

pub fn renewal_summary(self, input: RenewalSummary) -> Self

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

pub fn set_renewal_summary(self, input: Option<RenewalSummary>) -> Self

pub fn key_usages(self, input: impl Into<KeyUsage>) -> Self

pub fn set_key_usages(self, input: Option<Vec<KeyUsage>>) -> Self

pub fn extended_key_usages(self, input: impl Into<ExtendedKeyUsage>) -> Self

pub fn set_extended_key_usages(
self,
input: Option<Vec<ExtendedKeyUsage>>
) -> Self

pub fn certificate_authority_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

pub fn set_certificate_authority_arn(self, input: Option<String>) -> Self

pub fn renewal_eligibility(self, input: RenewalEligibility) -> Self

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

pub fn set_renewal_eligibility(self, input: Option<RenewalEligibility>) -> Self

pub fn options(self, input: CertificateOptions) -> Self

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

pub fn set_options(self, input: Option<CertificateOptions>) -> Self

pub fn build(self) -> CertificateDetail

Consumes the builder and constructs a CertificateDetail