Module aws_sdk_acm::operation

Structs

AddTagsToCertificate

Adds one or more tags to an ACM certificate. Tags are labels that you can use to

DeleteCertificate

Deletes a certificate and its associated private key. If this action succeeds, the

DescribeCertificate

Returns detailed metadata about the specified ACM certificate.

ExportCertificate

Exports a private certificate issued by a private certificate authority (CA) for use

GetAccountConfiguration

Returns the account

GetCertificate

Retrieves an Amazon-issued certificate and its certificate chain. The chain consists of

ImportCertificate

Imports a certificate into Amazon Web Services Certificate Manager (ACM) to use with services that are integrated with

ListCertificates

Retrieves a list of certificate ARNs and domain names. You can request that only

ListTagsForCertificate

Lists the tags that have been applied to the ACM certificate. Use the certificate's

PutAccountConfiguration

Adds or modifies

RemoveTagsFromCertificate

Remove one or more tags from an ACM certificate. A tag consists of a key-value pair. If

RenewCertificate

Renews an eligible ACM certificate. At this time, only exported private certificates can

RequestCertificate

Requests an ACM certificate for use with other Amazon Web Services services. To request an ACM

ResendValidationEmail

Resends the email that requests domain ownership validation. The domain owner or an

UpdateCertificateOptions

Updates a certificate. Currently, you can use this function to specify whether to opt in